Moved the home screens into the home package
