+++ app/src/main/java/com/betocrod/songrecords/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.betocrod.songrecords.ui.theme.SongRecordsTheme
+import com.betocrod.designsystem.SongRecordsTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
