--- app/src/main/java/com/betocrod/songrecords/domain/models/Song.kt
@@ -1,8 +1,0 @@
-package com.betocrod.songrecords.domain.models
-
-data class Song(
-    val name: String,
-    val image: String,
-    val artist: String,
-    val year: Int
-)
