+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,3 @@
 <resources>
     <string name="app_name">Song Records</string>
-    <string name="feature_home_no_songs_placeholder">There are no songs</string>
-    <string name="feature_home_error">Unexpected error</string>
-    <string name="feature_home_title">Song Records</string>
 </resources>
