+++ designSystem/src/main/java/com/betocrod/designsystem/TypeAlias.kt
@@ -1,0 +1,3 @@
+package com.betocrod.designsystem
+
+typealias DSDrawable = R.drawable
