+++ designSystem/src/main/res/drawable/ic_image.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M9,42q-1.2,0 -2.1,-0.9Q6,40.2 6,39L6,9q0,-1.2 0.9,-2.1Q7.8,6 9,6h30q1.2,0 2.1,0.9 0.9,0.9 0.9,2.1v30q0,1.2 -0.9,2.1 -0.9,0.9 -2.1,0.9ZM9,39h30L39,9L9,9v30ZM11.8,34.15h24.45l-7.35,-9.8 -6.6,8.55 -4.65,-6.35ZM9,39L9,9v30Z"/>
+</vector>
