+++ features/home/impl/src/main/java/com/betocrod/features/home/impl/domain/models/Song.kt
@@ -1,0 +1,8 @@
+package com.betocrod.features.home.impl.domain.models
+
+data class Song(
+    val name: String,
+    val image: String,
+    val artist: String,
+    val year: Int
+)
