+++ features/home/impl/src/main/java/com/betocrod/features/home/impl/widgets/previewparameters/SampleSongProvider.kt
@@ -1,0 +1,28 @@
+package com.betocrod.features.home.impl.widgets.previewparameters
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import com.betocrod.features.home.impl.domain.models.Song
+
+class SampleSongProvider : PreviewParameterProvider<Song> {
+    override val values: Sequence<Song>
+        get() {
+            return getSongList().asSequence()
+        }
+}
+
+class SampleSongListProvider : PreviewParameterProvider<List<Song>> {
+    override val values: Sequence<List<Song>>
+        get() = sequenceOf(
+            getSongList(),
+            emptyList()
+        )
+}
+
+internal fun getSongList() = (1..10).map {
+    Song(
+        name = "Song name #$it",
+        image = "https://www.imageurl.com",
+        artist = "Artist #$it",
+        year = 2001
+    )
+}
